Crate connector_agent

Re-exports

pub use crate::data_sources::csv::CSVSource;

pub use crate::data_sources::csv::CSVSourceBuilder;

pub use crate::data_sources::mixed::MixedSource;

pub use crate::data_sources::mixed::MixedSourceBuilder;

pub use crate::data_sources::postgres::PostgresDataSource;

pub use crate::data_sources::postgres::PostgresDataSourceBuilder;

pub use crate::data_sources::DataSource;

pub use crate::data_sources::SourceBuilder;

pub use crate::writers::Consume;

pub use crate::writers::PartitionWriter;

pub use crate::writers::Writer;

Modules

data_sources

writers

Structs

AnyArray

AnyArrayView

AnyArrayViewMut

Dispatcher
A dispatcher owns a SourceBuilder SB and a vector of queries schema is a temporary input before we implement infer schema or get schema from DB.

Enums

ConnectorAgentError
Errors that can be raised from this library.

DataOrder

DataType
This is our intermediate type system used in this library. For all the sources, their output values must be one of the types defined by DataType. For all the writers, they must support writing any value whose type is defined by DataType.

Traits

ParameterizedFunc
A ParameterizedFunc refers to a function that is parameterized on a type T, where type T will be dynaically determined by the variant of a TypeSystem. An example is the transmit<S,W,T> function. When piping values from a source to the writer, its type T is determined by the schema at the runtime.

ParameterizedOn
ParameterizedOn indicates a parameterized function Self is parameterized on type T

Realize
Realize means that a TypeSystem can realize a parameterized func F, based on its current variants.

TypeAssoc
Associate a static type to a TypeSystem

TypeSystem
TypeSystem describes a type system in a value type (e.g. enum variants), which can be used to type check with a static type T through the check method.

Type Definitions

Result