Module connector_agent::data_sources::dummy

Structs

BoolCounterSource
This DataSource only produces T which can be derived from bool.

BoolSourceBuilder

F64CounterSource
This DataSource only produces T which can be derived from f64.

F64SourceBuilder

OptU64SourceBuilder

OptU64TestSource

StringSource
This DataSource only produces T which can be derived from String

StringSourceBuilder

U64CounterSource
This DataSource only produces T which can be derived from u64.

U64SourceBuilder