Trait connector_agent::data_sources::DataSource

pub trait DataSource: Sized {
    type TypeSystem: TypeSystem;
    fn prepare(&mut self, query: &str) -> Result<()>;

    fn nrows(&self) -> usize;

    fn ncols(&self) -> usize;

    fn read<T>(&mut self) -> Result<T>
    where
        T: TypeAssoc<Self::TypeSystem>,
        Self: Produce<T>,
    { ... }

    fn infer_schema(&mut self) -> Result<Vec<Self::TypeSystem>> { ... }
}

In general, a DataSource abstracts the data source as a stream, which can produce a sequence of values of variate types by repetitively calling the function produce.

Associated Types

type TypeSystem: TypeSystem

The type system this DataSource associated with.

Loading content...
Required methods

fn prepare(&mut self, query: &str) -> Result<()>

Run the query and put the result into Self.

fn nrows(&self) -> usize

Number of rows this DataSource got.

fn ncols(&self) -> usize

Number of cols this DataSource got.

Loading content...
Provided methods

fn read<T>(&mut self) -> Result<T> where
T: TypeAssoc<Self::TypeSystem>,
Self: Produce<T>,

Read a value T by calling Produce<T>::produce. Usually this function does not need to be implemented.

fn infer_schema(&mut self) -> Result<Vec<Self::TypeSystem>>

Infer schema from return data

Loading content...