Macro connector_agent::associate_typesystems

macro_rules! associate_typesystems {
    (($ts1:ty, $ts2:ty), $($(([$($v1:tt)+], [$($v2:tt)+]))|+ => ($t1:ty, $t2:ty) conversion $cast:ident,)+) => { ... };
    (Realize ($ts1:ty, $ts2:ty), $($(($v1:pat, $v2:pat))|+ => ($t1:ty, $t2:ty),)+) => { ... };
    (Conversion ($ts1:ty, $ts2:ty), $(($v1:pat, $v2:expr),)+) => { ... };
    (TypeConversion all $ts1:ty, $ts2:ty, $t1:ty, $t2:ty) => { ... };
    (TypeConversion half $ts1:ty, $ts2:ty, $t1:ty, $t2:ty) => { ... };
    (TypeConversion none $ts1:ty, $ts2:ty, $t1:ty, $t2:ty) => { ... };
}

A macro to define how to convert between one type system to another, in terms of both their enum variants and the physical types

Example Usage

associate_typesystems! {
    (PostgresDTypes, DataType),
    ([PostgresDTypes::Float4], [DataType::F64]) => (f32, f64) conversion all,
}

This means for the type system PostgresDTypes, is can be converted to another type system DataType. Specifically, the variant PostgresDTypes::Float4(false) is corresponding to the variant DataType::F64(false) and the variant PostgresDTypes::Float4(true) is corresponding to the variant DataType::F64(true). Also, the physical type f32 is corresponding to the physical type f64 and the physical type Option<f32> is corresponding to the type Option<f64>. The last piece of conversion all means auto implementing the physical type conversion trait TypeConversion<f32, f64> and conversion trait TypeConversion<Option<f32>, Option<f64>> for (PostgresDTypes, DataType) by using casting rule (v as f64). If this is set to half, you need to manually write an TypeConversion implementation for (f32,f64), but the conversion rule for (Option<f32>, Option<f64>) is still generated. conversion none means generate nothing.