Trait connector_agent::sources::SourcePartition

pub trait SourcePartition: Sized {
    type TypeSystem: TypeSystem;
    type Parser: PartitionParser<'a, TypeSystem = Self::TypeSystem>;
    fn prepare(&mut self) -> Result<()>;

    fn parser(&mut self) -> Result<Self::Parser>;

    fn nrows(&self) -> usize;

    fn ncols(&self) -> usize;
}

In general, a DataSource abstracts the data source as a stream, which can produce a sequence of values of variate types by repetitively calling the function produce.

Associated Types

type TypeSystem: TypeSystem

type Parser: PartitionParser<'a, TypeSystem = Self::TypeSystem>

Loading content...
Required methods

fn prepare(&mut self) -> Result<()>

Run the query and put the result into Self.

fn parser(&mut self) -> Result<Self::Parser>

fn nrows(&self) -> usize

Number of rows this DataSource got. Sometimes it is not possible for the source to know how many rows it gets before reading the whole data.

fn ncols(&self) -> usize

Number of cols this DataSource got.

Loading content...