Trait connector_agent::writers::PartitionWriter

pub trait PartitionWriter<'a>: Send {
    type TypeSystem: TypeSystem;
    fn nrows(&self) -> usize;

    fn ncols(&self) -> usize;

    unsafe fn write<'r, T>(&'r mut self, value: T)
    where
        T: TypeAssoc<Self::TypeSystem> + 'r,
        Self: Consume<T>,
    { ... }

    fn write_checked<T>(&mut self, value: T) -> Result<()>
    where
        T: TypeAssoc<Self::TypeSystem>,
        Self: Consume<T>,
    { ... }

    fn finalize(&mut self) -> Result<()> { ... }
}

PartitionWriter writes values to its own region. PartitionWriter is parameterized on lifetime 'a, which is the lifetime of the parent Writer. This indicates the PartitionWriter can never live longer than the parent.

Associated Types

type TypeSystem: TypeSystem

Loading content...
Required methods

fn nrows(&self) -> usize

Number of rows this PartitionWriter controls.

fn ncols(&self) -> usize

Number of rows this PartitionWriter controls.

Loading content...
Provided methods

unsafe fn write<'r, T>(&'r mut self, value: T) where
T: TypeAssoc<Self::TypeSystem> + 'r,
Self: Consume<T>,

Write a value of type T to the location (row, col). The value is unchecked against the schema. This function is unsafe due to unchecked.

fn write_checked<T>(&mut self, value: T) -> Result<()> where
T: TypeAssoc<Self::TypeSystem>,
Self: Consume<T>,

Write a value of type T to the location (row, col), checked version. If T mismatch with the schema, ConnectorAgentError::UnexpectedType will return.

fn finalize(&mut self) -> Result<()>

Final clean ups

Loading content...