Trait connector_agent::writers::Writer

pub trait Writer<'a>: Sized {
    type TypeSystem;
    type PartitionWriter: PartitionWriter<'a, TypeSystem = Self::TypeSystem>;
    fn allocate(
        nrow: usize, 
        schema: Vec<Self::TypeSystem>
    ) -> Result<Self, ConnectorAgentError>;

    fn partition_writers(
        &'a mut self, 
        counts: &[usize]
    ) -> Vec<Self::PartitionWriter>;

    fn schema(&self) -> &[Self::TypeSystem];
}

A Writer is associated with a TypeSystem and a PartitionWriter. PartitionWriter allows multiple threads write data into the buffer owned by Writer.

Associated Types

type TypeSystem

type PartitionWriter: PartitionWriter<'a, TypeSystem = Self::TypeSystem>

Loading content...
Required methods

fn allocate(
nrow: usize,
schema: Vec<Self::TypeSystem>
) -> Result<Self, ConnectorAgentError>

Construct the Writer. This allocates the memory based on the types of each columns and the number of rows.

fn partition_writers(
&'a mut self,
counts: &[usize]
) -> Vec<Self::PartitionWriter>

Create a bunch of partition writers, with each write count number of rows.

fn schema(&self) -> &[Self::TypeSystem]

Return the schema of the writer.

Loading content...