Crate connector_agent

Re-exports

pub use crate::data_sources::csv::CSVSource;

pub use crate::data_sources::dummy::BoolCounterSource;

pub use crate::data_sources::dummy::StringSource;

pub use crate::data_sources::dummy::U64CounterSource;

pub use crate::writers::dummy::BoolPartitionWriter;

pub use crate::writers::dummy::BoolWriter;

pub use crate::writers::dummy::StringPartitionWriter;

pub use crate::writers::dummy::StringWriter;

pub use crate::writers::dummy::U64PartitionWriter;

pub use crate::writers::dummy::U64Writer;

pub use crate::writers::PartitionWriter;

pub use crate::writers::Writer;

Modules

data_sources

writers

Structs

Worker
A worker owns a DataSource S and a PartitionWriter W. It is also parameterized on a TypeSystem TS.

WriteWorker
A worker owns a DataSource S and a PartitionWriter W. It is also parameterized on a TypeSystem TS.

Enums

ConnectorAgentError
Errors that can be raised from this library.

DataType
This is our intermediate type system used in this library. For all the sources, their output values must be one of the types defined by DataType. For all the writers, they must support writing any value whose type is defined by DataType.

Traits

Transmit
Transmit defines Self that can pull data from a data source S and push the data to the writer W.

TypeSystem
TypeSystem<T> maps every type in Self to a concrete native type T. Usually multiple TypeSystem<T> will be implemented for a same Self type.