Trait connector_agent::data_sources::DataSource

pub trait DataSource {
    type TypeSystem;
    fn run_query(&mut self, query: &str) -> Result<(), ConnectorAgentError>;

    fn produce<T>(&mut self) -> Result<T, ConnectorAgentError>
    where
        Self::TypeSystem: TypeSystem<T>,
        Self: Parse<T>,
    { ... }
}

A DataSource should be able to run_query and store the query result in its own buffer. A DataSource should also be able to produce any type T, which is defined by its associated TypeSystem, by calling the function DataSource::produce.

Associated Types

type TypeSystem

The type system this DataSource associated with.

Loading content...
Required methods

fn run_query(&mut self, query: &str) -> Result<(), ConnectorAgentError>

Run the query and put the result into Self.

Loading content...
Provided methods

fn produce<T>(&mut self) -> Result<T, ConnectorAgentError> where
Self::TypeSystem: TypeSystem<T>,
Self: Parse<T>,

Produce a value T by calling Parse<T>::parse. Usually this function does not need to be implemented.

Loading content...