Trait connector_agent::writers::PartitionWriter

pub trait PartitionWriter<'a> {
    type TypeSystem;
    unsafe fn write<T>(&mut self, row: usize, col: usize, value: T);

    fn write_checked<T>(
        &mut self, 
        row: usize, 
        col: usize, 
        value: T
    ) -> Result<(), ConnectorAgentError>
    where
        Self::TypeSystem: TypeSystem<T>;

    fn nrows(&self) -> usize;

    fn ncols(&self) -> usize;
}

PartitionWriter writes values to its own region. PartitionWriter is parameterized on lifetime 'a, which is the lifetime of the parent Writer. This indicates the PartitionWriter can never live longer than the parent.

Associated Types

type TypeSystem

Loading content...
Required methods

unsafe fn write<T>(&mut self, row: usize, col: usize, value: T)

Write a value of type T to the location (row, col). The value is unchecked against the schema. This function is unsafe due to unchecked.

fn write_checked<T>(
&mut self,
row: usize,
col: usize,
value: T
) -> Result<(), ConnectorAgentError> where
Self::TypeSystem: TypeSystem<T>,

Write a value of type T to the location (row, col), checked version. If T mismatch with the schema, ConnectorAgentError::UnexpectedType will return.

fn nrows(&self) -> usize

Number of rows this PartitionWriter controls.

fn ncols(&self) -> usize

Number of rows this PartitionWriter controls.

Loading content...