Trait connector_agent::writers::Writer

pub trait Writer: Sized {
    type TypeSystem: TypeSystem;
    type PartitionWriter: PartitionWriter<'a, TypeSystem = Self::TypeSystem>;

    const DATA_ORDERS: &'static [DataOrder];

    fn allocate<S: AsRef<str>>(
        &mut self, 
        nrow: usize, 
        names: &[S], 
        schema: &[Self::TypeSystem], 
        data_order: DataOrder
    ) -> Result<()>;

    fn partition_writers(
        &mut self, 
        counts: &[usize]
    ) -> Result<Vec<Self::PartitionWriter>>;

    fn schema(&self) -> &[Self::TypeSystem];
}

A Writer is associated with a TypeSystem and a PartitionWriter. PartitionWriter allows multiple threads write data into the buffer owned by Writer.

Associated Types

type TypeSystem: TypeSystem

type PartitionWriter: PartitionWriter<'a, TypeSystem = Self::TypeSystem>

Loading content...
Associated Constants

const DATA_ORDERS: &'static [DataOrder]

Loading content...
Required methods

fn allocate<S: AsRef<str>>(
&mut self,
nrow: usize,
names: &[S],
schema: &[Self::TypeSystem],
data_order: DataOrder
) -> Result<()>

Construct the Writer. This allocates the memory based on the types of each columns and the number of rows.

fn partition_writers(
&mut self,
counts: &[usize]
) -> Result<Vec<Self::PartitionWriter>>

Create a bunch of partition writers, with each write count number of rows.

fn schema(&self) -> &[Self::TypeSystem]

Return the schema of the writer.

Loading content...